fix: rename files to follow convention
